+++ src/main/java/bots/Bot.java
@@ -1,0 +1,5 @@
+package bots;
+
+public class Bot {
+
+}
