+++ src/main/java/bots/BotManager.java
@@ -1,0 +1,5 @@
+package bots;
+
+public class BotManager implements IBotManager {
+
+}
