+++ src/main/java/bots/IBotManager.java
@@ -1,0 +1,5 @@
+package bots;
+
+public interface IBotManager {
+
+}
