+++ src/main/java/gsheetsapi/GSheetsAdapter.java
@@ -1,0 +1,5 @@
+package gsheetsapi;
+
+public class GSheetsAdapter implements IGSheetsAdapter {
+
+}
