+++ src/main/java/gsheetsapi/IGSheetsAdapter.java
@@ -1,0 +1,5 @@
+package gsheetsapi;
+
+public interface IGSheetsAdapter {
+
+}
