+++ src/main/java/instaapi/RIInstaAPI.java
@@ -1,0 +1,24 @@
+package instaapi;
+
+public class RIInstaAPI implements IInstaAdapter {
+
+  @Override
+  public void login(String username, String password) {
+
+  }
+
+  @Override
+  public void logout() {
+
+  }
+
+  @Override
+  public void followUser(int accountId) {
+
+  }
+
+  @Override
+  public void likePost(int postId) {
+
+  }
+}
