+++ src/main/java/main/MainApp.java
@@ -1,4 +1,4 @@
-package view;
+package main;
 
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
