+++ src/main/resources/main/scene.fxml
@@ -9,7 +9,7 @@
 <StackPane xmlns:fx="http://javafx.com/fxml/1" maxHeight="-Infinity" maxWidth="-Infinity"
   minHeight="-Infinity"
   minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0"
-  xmlns="http://javafx.com/javafx/8" fx:controller="view.AppController">
+  xmlns="http://javafx.com/javafx/8" fx:controller="main.AppController">
   <BorderPane prefHeight="200.0" prefWidth="200.0">
     <top>
       <HBox BorderPane.alignment="CENTER" prefHeight="100.0" prefWidth="200.0">
